+++ mediabus-rtp-demo/rtp-sender/send-pcm16.sh
@@ -1,0 +1,19 @@
+#!/usr/bin/env bash
+
+HERE=$(realpath $(dirname "$0"))
+
+PORT=${1?port missing}
+MY_IP=${2?host ip missing}
+FNAME=${3:-${HERE}/28797-04.ogg}
+FILE=$(realpath $FNAME)
+
+gst-launch-1.0  uridecodebin uri=file://$FILE ! \
+                queue ! \
+                audioconvert ! \
+                audioresample ! \
+                audio/x-raw,format=S16BE,rate=16000,channels=1 ! \
+                rtpL16pay pt=100 \
+                      min-ptime=10000000 \
+                      max-ptime=200000000 \
+                      ptime-multiple=10000000 ! \
+                udpsink host=$MY_IP port=$PORT