+++ shell.nix
@@ -1,29 +1,10 @@
-{ pkgs ? (import ./nix/pkgs.nix) }:
-let
-  this = pkgs.callPackage ./default.nix {};
-in
-this.shellFor {
-  packages = ps: [ ps.mediabus-rtp ];
-  buildInputs = with pkgs.gst_all_1;
-    [
-      gstreamer
-      gst-plugins-base
-      gst-plugins-good
-      gst-plugins-bad
-      pkgs.sox
-    ];
-
-  # HACK
-  # make hspec-discover available
-  # https://github.com/input-output-hk/haskell.nix/issues/231#issuecomment-601469249
-  exactDeps = false;
-  # END OF HACK
-
-  withHoogle = true;
-  tools = {
-    cabal = "latest";
-    ormolu = "latest";
-    haskell-language-server = "latest";
-    hlint = "latest";
-  };
-}
+(import
+  (
+    let lock = builtins.fromJSON (builtins.readFile ./flake.lock); in
+    fetchTarball {
+      url = "https://github.com/edolstra/flake-compat/archive/${lock.nodes.flake-compat.locked.rev}.tar.gz";
+      sha256 = lock.nodes.flake-compat.locked.narHash;
+    }
+  )
+  { src = ./.; }
+).shellNix